+++ Assets/HandPainted Lava Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 36d0a6ead542add40bb6d774a0fe14da
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
