+++ Assets/HandPainted Lava Textures/Demo/Scene.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a0cf9f1096220274388d707860ce81fe
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
