+++ Assets/HandPainted Lava Textures/Documentation/Documentation.txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: afbf2642bc01bd94fb2d203970931425
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
