+++ Assets/HandPainted Lava Textures/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e96c669152a3f7c44bae1633ca09a197
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
