+++ Assets/HandPainted Lava Textures/Materials/01.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c262c90b10ca8734f9c612ad941e1220
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
