+++ Assets/HandPainted Lava Textures/Materials/02.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 745b04538c7c0a740b168e1c106307ac
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
