+++ Assets/HandPainted Lava Textures/Materials/03.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7d02d22da273a1d4fa60e5a5350b9a16
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
