+++ Assets/HandPainted Lava Textures/Materials/04.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e054df7fe20dcd241b757ce007d6674f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
