+++ Assets/HandPainted Lava Textures/Materials/1.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a899cbe0b809ecb4594bf852d7f00ce0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
