+++ Assets/HandPainted Lava Textures/Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: efcc25a8125262a4bb3b00d56013b197
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
