+++ Assets/Modelado3D/Lava_Flowing_Shader/Materials/lava1.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0d8578ef7529532459b4cd68021e94dc
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
