+++ Assets/Modelado3D/Lava_Flowing_Shader/Materials/lava3.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0548120a020d4a94786ca1f63c5f4bff
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
