--- Assets/Scripts/Dona.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e9cbba0ce527dda4594534965e6b42b5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
