+++ UserSettings/EditorUserSettings.asset
@@ -15,10 +15,10 @@ EditorUserSettings:
       value: 575301575c020d090c5d5b2140745c4413164e737c7e71602c7b4832b0b8663c
       flags: 0
     RecentlyUsedSceneGuid-3:
-      value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
+      value: 5257065751530a0c0c5b0f2714710744174e4b28282d23607c2b4b64b1b83268
       flags: 0
     RecentlyUsedSceneGuid-4:
-      value: 5257065751530a0c0c5b0f2714710744174e4b28282d23607c2b4b64b1b83268
+      value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
